server/requirements.txt
certifi>=2018.1.18
chardet>=3.0.4
click>=6.7
Flask>=1.1.1
idna>=2.6
itsdangerous>=0.24
Jinja2>=2.11.3
MarkupSafe>=1.1.1
requests>=2.22.0
themis.finals.attack.helper>=1.1.0
themis.finals.attack.result>=1.3.0
urllib3>=1.26.5
Werkzeug>=0.16.0
psycopg2-binary
